requirements.txt
branca==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
folium==0.15.0
fonttools==4.46.0
idna==3.6
importlib-resources==6.1.1
isodate==0.6.1
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.6
importlib-metadata==6.9.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
